+++ Plugins/DynamicRef/Source/NexusDynamicRefEditor/Private/NDynamicRefEditorStyle.cpp
@@ -1,0 +1,21 @@
+﻿// Copyright dotBunny Inc. All Rights Reserved.
+// See the LICENSE file at the repository root for more information.
+
+#include "NDynamicRefEditorStyle.h"
+
+#include "Brushes/SlateImageBrush.h"
+#include "Macros/NEditorStyleMacros.h"
+#include "Styling/SlateStyle.h"
+
+N_IMPLEMENT_EDITOR_STYLE(FNDynamicRefEditorStyle)
+
+TSharedRef<FSlateStyleSet> FNDynamicRefEditorStyle::Create()
+{
+	N_IMPLEMENT_EDITOR_STYLE_CREATE
+	
+	// NDynamicRefComponent
+	Style.Set("ClassIcon.NDynamicRefComponent", new N_MODULE_IMAGE_BRUSH_SVG(PluginDirectory, TEXT("ClassIcon_NDynamicRefComponent"), Icon16x16));
+	Style.Set("ClassThumbnail.NDynamicRefComponent",  new N_MODULE_IMAGE_BRUSH_SVG(PluginDirectory, TEXT("ClassIcon_NDynamicRefComponent"), Icon128x128));
+
+	return StyleRef;
+}